Move genUid back.
