+++ formats/keep/util.ts
@@ -34,13 +34,6 @@ export function sanitizeTags(str: string): string {
 	return newStr;
 }
 
-export function genUid(length: number): string {
-	let array: string[] = [];
-	for (let i = 0; i < length; i++) {
-		array.push((Math.random() * 16 | 0).toString(16));
-	}
-	return array.join('');
-}
 /**
  * Takes a string and returns in lowercase with the first letter capitalised.
  */
