+++ formats/yarle/utils/turndown-rules/internal-links-rule.ts
@@ -1,4 +1,4 @@
-import { genUid } from "formats/keep/util";
+import { genUid } from '../../../../util';
 import { RuntimePropertiesSingleton } from '../../runtime-properties';
 import { yarleOptions } from '../../yarle';
